Add files via upload
This commit is an update for creating the dataset. There have been multiple roadblocks found in creating this dataset so far, to be fixed by the next commit. Also in the next commit will be standard Exploratory Data Analysis, and a more detailed insight on the final goal of this project.